Fix: Mobile navigation
Fixes mobile navigation by refactoring what was in the `index.tsx` for the sidebar into the respective
mobile and desktop components.
