+++ src/theme/DocSidebar/index.tsx
@@ -1,87 +1,23 @@
-import React from 'react';
-import { useLocation } from '@docusaurus/router';
-import type { PropSidebarItem } from '@docusaurus/plugin-content-docs';
-import Link from '@docusaurus/Link';
-import {
-  buildCategories,
-  isActiveLink,
-  hasActiveChild,
-  type Category,
-} from './utils';
-import './CustomSidebar.css';
+import React, {type ReactNode} from 'react';
+import {useWindowSize} from '@docusaurus/theme-common';
+import DocSidebarDesktop from '@theme/DocSidebar/Desktop';
+import DocSidebarMobile from '@theme/DocSidebar/Mobile';
+import type {Props} from '@theme/DocSidebar';
 
-interface CustomSidebarProps {
-  sidebar: PropSidebarItem[];
-}
+export default function DocSidebar(props: Props): ReactNode {
+  const windowSize = useWindowSize();
 
-const CustomSidebar: React.FC<CustomSidebarProps> = ({ sidebar }) => {
-  const location = useLocation();
-  const categories: Category[] = buildCategories(sidebar);
+  // Desktop sidebar visible on hydration: need SSR rendering
+  const shouldRenderSidebarDesktop =
+    windowSize === 'desktop' || windowSize === 'ssr';
 
-  const renderSidebarItem = (item: PropSidebarItem): React.ReactElement => {
-    if (item.type === 'link') {
-      const href = (item as any).href || '';
-      const isActive = isActiveLink(href, location.pathname);
-
-      return (
-        <Link
-          to={href}
-          className={`custom-sidebar__link ${isActive ? 'custom-sidebar__link--active' : ''}`}
-        >
-          {(item as any).label || 'Untitled'}
-        </Link>
-      );
-    }
-
-    if (item.type === 'category') {
-      const items = (item as any).items;
-      if (items && Array.isArray(items)) {
-        const shouldExpand = hasActiveChild(items, location.pathname);
-
-        return (
-          <details className="custom-sidebar__collapsible" open={shouldExpand}>
-            <summary className="custom-sidebar__folder-title">
-              {(item as any).label || 'Untitled Category'}
-            </summary>
-            <ul className="custom-sidebar__subitems">
-              {items.map((subItem: PropSidebarItem, subIndex: number) => (
-                <li key={subIndex}>{renderSidebarItem(subItem)}</li>
-              ))}
-            </ul>
-          </details>
-        );
-      }
-    }
-
-    return <span>{(item as any).label || 'Unknown item'}</span>;
-  };
-
-  const renderCategory = (category: Category) => (
-    <div key={category.id} className="custom-sidebar__category">
-      <h3 className="custom-sidebar__category-title">{category.title}</h3>
-      <div className="custom-sidebar__category-content">
-        {category.items.length === 0 ? (
-          <p className="custom-sidebar__empty">No items yet</p>
-        ) : (
-          <div className="custom-sidebar__items">
-            {category.items.map((item, index) => (
-              <div key={index} className="custom-sidebar__item">
-                {renderSidebarItem(item)}
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-    </div>
-  );
+  // Mobile sidebar not visible on hydration: can avoid SSR rendering
+  const shouldRenderSidebarMobile = windowSize === 'mobile';
 
   return (
-    <div className="custom-sidebar">
-      <div className="custom-sidebar__content">
-        {categories.map(renderCategory)}
-      </div>
-    </div>
+    <>
+      {shouldRenderSidebarDesktop && <DocSidebarDesktop {...props} />}
+      {shouldRenderSidebarMobile && <DocSidebarMobile {...props} />}
+    </>
   );
-};
-
-export default CustomSidebar;
+}
